fix: resolve integer overflow and redundant shuffle issues
Resolves review comments on PR #326:
- UniformEpisodicDataLoaderTests.cs:32 - convert integers to double before multiplication to prevent overflow
- UniformEpisodicDataLoader.cs:187 - remove redundant second shuffle (already shuffled during sampling)

+++ tests/UnitTests/Data/UniformEpisodicDataLoaderTests.cs
@@ -29,7 +29,7 @@ public class UniformEpisodicDataLoaderTests
                 // Populate features with unique values based on class and example
                 for (int featureIdx = 0; featureIdx < numFeatures; featureIdx++)
                 {
-                    X[rowIdx, featureIdx] = classIdx * 1000 + exampleIdx * 10 + featureIdx;
+                    X[rowIdx, featureIdx] = (double)classIdx * 1000.0 + (double)exampleIdx * 10.0 + (double)featureIdx;
                 }
 
                 // Set label
